made the decorator generic
