+++ src/snapred/backend/data/LocalDataService.py
@@ -30,8 +30,9 @@ from snapred.backend.dao.state import (
 )
 from snapred.backend.dao.state.CalibrantSample import CalibrantSamples
 from snapred.backend.data.LocalWorkspaceDataService import LocalWorkspaceDataService
+from snapred.backend.error.StateValidationException import StateValidationException
 from snapred.meta.Config import Config, Resource
-from snapred.meta.decorators.ErrorHandler import StateExceptionHandler
+from snapred.meta.decorators.ExceptionHandler import ExceptionHandler
 from snapred.meta.decorators.Singleton import Singleton
 from snapred.meta.mantid.WorkspaceNameGenerator import WorkspaceName
 from snapred.meta.redantic import (
@@ -204,7 +205,7 @@ class LocalDataService:
             raise FileNotFoundError("File {} does not exist".format(fName))
         return f
 
-    @StateExceptionHandler
+    @ExceptionHandler(StateValidationException)
     def _generateStateId(self, runId: str) -> Tuple[Any, Any]:
         if runId in self.stateIdCache:
             return self.stateIdCache[runId]
@@ -557,7 +558,7 @@ class LocalDataService:
         # write the file and return the calibration state
         write_model_pretty(calibration, calibrationParametersPath)
 
-    @StateExceptionHandler
+    @ExceptionHandler(StateValidationException)
     def initializeState(self, runId: str, name: str = None):
         # pull pv data similar to how we generate stateId
         pvFile = self._readPVFile(runId)
